ocr-engine/requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0